feat: replace existing volume rendering example
