--- packages/core/examples/volume_rendering_multichannel/index.html
@@ -1,37 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Volume Rendering Multichannel</title>
-    <style>
-      html,
-      body {
-        margin: 0;
-        padding: 0;
-        display: flex;
-        width: 100%;
-        height: 100%;
-        background-color: black;
-        font-family: monospace;
-        font-size: 12px;
-      }
-      #main {
-        display: flex;
-        flex-direction: column;
-        width: 100%;
-        height: 100%;
-      }
-      #canvas {
-        width: 100%;
-        height: 100%;
-      }
-    </style>
-  </head>
-  <body>
-    <div id="main">
-      <canvas id="canvas"></canvas>
-    </div>
-    <script type="module" src="./main.ts"></script>
-  </body>
-</html>
